hal9001-backend/requirements.txt
# Web server
fastapi
uvicorn[standard]

# PostgreSQL database driver (async)
psycopg[binary,pool]

# For reading .env files
python-dotenv